faust-processor/requirements.txt
faust-streaming==0.11.3
pydantic==2.9.0
pydantic-settings==2.4.0
python-schema-registry-client[faust]